Granular GC handling; bactraces; support for module importing. (#85)
Cleaned up GC code and made it so that, in most places, we only scan memory for pointers that represent fields with pointers. This basically is handled by per-datatype collection hooks, and did a bit of work on just the primary hatrack dictionary to support (per-datatype values that get passed through during record allocation).
Module importing is done, and the system __init.c4m is automatically imported (which currently imports one starting module).
Added C backtraces, including in a signal handler if there's a SEGV.
Small improvements to the test command line, such as searching your cwd for files if the file name isn't found in the test dir.
Got a bunch more tests working.
The pattern stuff is a bit broken due to changes; there are some work-arounds there. I don't plan to fight variadic macros right now, so the workarounds are going to stay for a while.

+++ include/compiler/datatypes/cfg.h
@@ -36,8 +36,8 @@ typedef struct {
     c4m_cfg_node_t  *exit_node;
     c4m_cfg_node_t **branch_targets;
     c4m_utf8_t      *label; // For loops
-    int32_t          num_branches;
-    int32_t          next_to_process;
+    int64_t          num_branches;
+    int64_t          next_to_process;
 } c4m_cfg_branch_info_t;
 
 typedef struct {
